requirements.txt
translate
translation
googletrans
kafka-python
langdetect
fuzzywuzzy
python-Levenshtein